// FrontEnd/src/app/models/Task.ts
export interface Task {
  taskName: string;
  description: string;
  courseName: string;
  taskType: string;
  dueDate: string;
  milestones: number;
  completionDate : Date;
  courseId: string;
  difficultyLevel : string
  notes : string;
  status : string
  taskId : string;
  title : string
  studentId : string;
  milestonesCompleted :  number;
}